requirements.txt
discord.py==0.16.3
beautifulsoup4==4.5.3
requests==2.12.4
